// payment-service/src/types/enum.types.ts
export enum environment_variables {
    DB_HOST = 'DB_HOST',
    DB_PORT = 'DB_PORT',
    DB_USERNAME = 'DB_USERNAME',
    DB_PASSWORD = 'DB_PASSWORD',
    DB_NAME = 'DB_NAME',
    RABBITMQ_URL = 'RABBITMQ_URL',
    RABBITMQ_EXCHANGE = 'RABBITMQ_EXCHANGE',
    AUTH_ROUTING_KEY = 'AUTH_ROUTING_KEY',
    RABBITMQ_DEFAULT_USER = 'RABBITMQ_DEFAULT_USER',
    RABBITMQ_DEFAULT_PASS = 'RABBITMQ_DEFAULT_PASS',
    FRONTEND_URL= 'FRONTEND_URL',
    SERVER_PORT= 'SERVER_PORT',
    NODE_ENV= 'NODE_ENV',
    AUTH_JWT_SECRET= 'AUTH_JWT_SECRET',
    GATEWAY_SECRET_KEY= 'GATEWAY_SECRET_KEY',
    PAYSTACK_SECRET_KEY= 'PAYSTACK_SECRET_KEY',
    PAYSTACK_BASE_URL= 'PAYSTACK_BASE_URL',
    REDIS_HOST='REDIS_HOST',
    REDIS_PORT='REDIS_PORT',
    REDIS_PASSWORD='REDIS_PASSWORD',
    REDIS_DB='REDIS_DB'
}


export enum payment_status {
    PENDING = 'pending',
    COMPLETED = 'completed',
    FAILED = 'failed',
    REFUNDED = 'refunded'
}